setup basic write and utils.
